Refactor cofidectl up functions into the spire package
This keeps the up command logic at a high level and allows us to reuse
some helper functions and constants in the spire package.

Fixes: #9
